deleting & updating only one's posts

+++ app/routers/post.py
@@ -22,7 +22,7 @@ def get_posts(db: Session=Depends(get_db),
 def create_post(post: schemas.PostCreate, 
                 db: Session=Depends(get_db),
                 current_user: int=Depends(oauth2.get_current_user)):
-    new_post=models.Post(**post.model_dump())
+    new_post=models.Post(owner_id=current_user.id,**post.model_dump())
     db.add(new_post) #adding to db
     db.commit()
     db.refresh(new_post) #refreshing the new post
@@ -59,6 +59,10 @@ def delete_post(id:int,
         raise  HTTPException(status_code=status.HTTP_404_NOT_FOUND, 
                              detail=f"Post with id {id} not found")
     
+    # check if user logged in is the owner of the post
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Not authorized to perform requested action")
     # Delete the post
     post_query.delete(synchronize_session=False)
     db.commit()
@@ -82,6 +86,11 @@ def update_post(id:int,
         raise  HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                              detail=f"Post with id {id} not found")
 
+    # check if user is the owner of the post
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Not authorized to perform requested action")
+    
     # Update the post
     post_query.update(updated_post.model_dump(), synchronize_session=False)
     db.commit()
